+++ cypress/support/accessibility/a11y-utils.js
@@ -54,6 +54,7 @@ export default (from, end) => {
       !prepareUrl[0].startsWith("alert") &&
       !prepareUrl[0].startsWith("action-popover") &&
       !prepareUrl[0].startsWith("anchor-navigation") &&
+      !prepareUrl[0].startsWith("loader") &&
       !prepareUrl[0].startsWith("loader-bar") &&
       !prepareUrl[0].startsWith("link-preview") &&
       !prepareUrl[0].startsWith("verticaldivider") &&
@@ -64,6 +65,9 @@ export default (from, end) => {
       !prepareUrl[0].startsWith("advanced-color-picker") &&
       !prepareUrl[0].startsWith("preview") &&
       !prepareUrl[0].startsWith("detail") &&
+      !prepareUrl[0].startsWith("help") &&
+      !prepareUrl[0].startsWith("toast") &&
+      !prepareUrl[0].startsWith("dialog-full-screen") &&
       !prepareUrl[0].endsWith("test")
     ) {
       urlList.push([prepareUrl[0], prepareUrl[1]]);